rename package util to perf
